Updated the samples to LocalStack 1.4
        The new samples now also carry the metadata inside the zip file.
        The pod content is now inside the folder containing the script
        to generate the pod's state.
