+++ happy-new-year/error.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <title>404</title>
+  </head>
+  <body>
+    <p>Something is amiss.</p>
+  </body>
+</html>
