+++ happy-new-year/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html><html lang="en"><head><meta http-equiv="Content-Type" content="text/html" /><meta  charset="utf-8"  /><title>Happy New Year</title></head><body style="width: 100%; text-align: center;"><div style="width: 800px; text-align: left; margin: 0px auto;"><img src="localstack-logo.png" alt="LocalStack"> 
+      <h2>Happy New Year to all of our valued users!</h2><p>
+        As we enter this new year, we at LocalStack would like to express our sincerest gratitude for your continued support.<br>
+        Your trust in our product means the world to us, and we are committed to delivering the best experience possible to you in the year ahead.
+      </p><p>
+        We have big plans for the coming year and can't wait to share them with you.<br>
+        In the meantime, we hope you have a fantastic start to the new year and that 2023 brings you all the happiness, success, and prosperity you deserve.
+      </p><p>
+        Thank you for being a part of the LocalStack community. We look forward to continuing to serve you in the year ahead.
+      </p>
+        
+        Wishing you all the best,<br>
+        The LocalStack Team</p><img style="width: 800px;" src="fireworks.jpg" alt="Fireworks"></div></body></html>
+
