--- thumbnail-s3-tutorial/Makefile
@@ -1,15 +1,0 @@
-export AWS_ACCESS_KEY_ID ?= test
-export AWS_SECRET_ACCESS_KEY ?= test
-export AWS_DEFAULT_REGION = us-east-1
-
-usage:      ## Show this help
-	@fgrep -h "##" $(MAKEFILE_LIST) | fgrep -v fgrep | sed -e 's/\\$$//' | sed -e 's/##//'
-
-install:    ## Install dependencies
-	@which localstack || pip install localstack
-	@which awslocal || pip install awscli-local
-
-create-state:        
-	@./script.sh
-
-.PHONY: usage install run test-ci
