--- thumbnail-s3-tutorial/configuration.json
@@ -1,11 +1,0 @@
-{
-    "LambdaFunctionConfigurations":[
-       {
-          "Id":"CreateThumbnail-s3-configuration",
-          "LambdaFunctionArn":"arn:aws:lambda:us-east-1:000000000000:function:CreateThumbnail",
-          "Events":[
-             "s3:ObjectCreated:*"
-          ]
-       }
-    ]
- }
